Add SyCamera and SyUV
SyUV distorts in UV space, and SyCamera renders 3D scenes with Syntheyes lens distortion baked in. This improves the quality dramatically while preserving filtering steps on the image.
Read the README.html on how to use the new nodes, and see the sample scene.
This thread on nuke-users is a good introduction to the workflow.
http://forums.thefoundry.co.uk/phpBB2/viewtopic.php?p=1915

This series of updates also speeds up the distortion computation using a lookup table, and lays the groundwork
for pluggable distortion. It also removes the "uncropping" of outputs since all of the output is preserved within
the bounding box, but outside the format area. This workflow is vastly superior to generating oversize plates because
you do not have to worry about centering, camera FOV changes and so on.