Set the right path for the stashed OSX 10.5 SDK
The stashed path is the "usr" directory from a running 10.5.8 machine, that includes GCC 4.0 which is ridiculously difficult to install on Lion. What we can do instead is just stash the whole "/usr" somewhere and build against that and the separate 10.5 SDK. Since library linker paths are relative, everything within "usr" stays working. The only thing we need to observe is that all the symlinks work properly.

So on Lion, to build plugins it is sufficient to stash your 10.5.8 "/urs" somewhere along with a 10.5 SDK header directory at the standard location, and put this "usr"'s "bin" onto your $PATH when compiling the plugins.